implementing edit functionality

+++ src/views/Home.vue
@@ -25,7 +25,7 @@
           <v-btn fab color="red" small @click="reject()">
             <v-icon color="white">mdi-close</v-icon>
           </v-btn>
-          
+
           <a id="check" @click="save()">indir</a>
         </div>
 
@@ -59,7 +59,8 @@ export default {
       output: "",
       isClicked: false,
       recordTime: "",
-      imgdata:[]
+      imgdata: [],
+      coordinates: {},
     };
   },
   methods: {
@@ -77,23 +78,36 @@ export default {
       // let isDrawing;
 
       canvasElement.onmousedown = (e) => {
+        console.log(e)
         context.beginPath();
-        context.closePath()
-if (e.offsetX) {
-          mouseX= e.offsetX;
-          mouseY=e.offsetY;
-        }else{
-          mouseX = e.pageX -e.target.offsetLeft;
+        context.closePath();
+        if (e.offsetX) {
+          mouseX = e.offsetX;
+          mouseY = e.offsetY;
+        } else {
+          mouseX = e.pageX - e.target.offsetLeft;
           mouseY = e.pageY - e.target.offsetTop;
         }
-        
-        canvasElement.onmouseup = function (e) {
-          context.strokeRect(mouseX, mouseY, e.offsetX-mouseX ,e.offsetY-mouseY);
+
+        let x = e.layerX;
+        let y = e.layerY
+
+        canvasElement.onmouseup = (e) => {
+          console.log(e)
+          this.coordinates.x1 = x;
+          this.coordinates.x2 = e.layerX;
+          this.coordinates.y1 = y;
+          this.coordinates.y2 = e.layerY;
+          context.strokeRect(
+            mouseX,
+            mouseY,
+            e.offsetX - mouseX,
+            e.offsetY - mouseY
+            
+          );
           context.closePath();
         };
       };
-
-      
     },
     async screenShot() {
       this.recordTime = new Date();
@@ -109,50 +123,50 @@ if (e.offsetX) {
       canvas.width = w;
       canvas.height = h;
 
-      console.log(w, h);
-
       context.fillRect(0, 0, w, h);
-      context.drawImage( img,0, 0, w, h);
-      console.log(canvas);
+      context.drawImage(img, 0, 0, w, h);
       this.drawOnImage();
     },
 
     save() {
       const canvas = document.getElementById("layer1");
-      canvas.crossOrigin = "Anonymous"
-      var dataURL = canvas.toDataURL("image/png").replace(/^data:image\/(png|jpg);base64,/, "");
+      canvas.crossOrigin = "Anonymous";
+      var dataURL = canvas
+        .toDataURL("image/png")
+        .replace(/^data:image\/(png|jpg);base64,/, "");
 
-      const data = {
-        url:dataURL,
-        timestamp:this.recordTime,
-        location:"",
+      let location = prompt("Please enter location name", "Turkey");
+      if (location != null) {
+        const data = {
+          url: dataURL,
+          timestamp: this.recordTime,
+          location: location,
+          coordinates: this.coordinates,
+        };
+
+        this.imgdata.push(data);
+        this.isClicked = false;
+        localStorage.setItem("imgData", JSON.stringify(this.imgdata));
       }
 
-      this.imgdata.push(data)
-
-      localStorage.setItem("imgData", JSON.stringify(this.imgdata));
-
       // var link = document.getElementById("check");
-      
-      
+
       //     link.href = canvas.toDataURL();
       //     link.download = "mypainting.png";
-        
     },
     reject() {
       this.isClicked = false;
     },
   },
-  created(){
-    let data = JSON.parse(localStorage.getItem("imgData"))
-    console.log(data)
-    if(data){
-      this.imgdata = data
-    }else{
-      this.imgdata=[]
+  created() {
+    let data = JSON.parse(localStorage.getItem("imgData"));
+    console.log(data);
+    if (data) {
+      this.imgdata = data;
+    } else {
+      this.imgdata = [];
     }
-    
-  }
+  },
 };
 </script>
 
@@ -180,7 +194,6 @@ if (e.offsetX) {
 }
 
 .popup {
-  
   border-radius: 10px;
   border-color: black;
   top: 0px;
